+++ renderer.py
@@ -1,0 +1,3 @@
+__author__ = 'charles.andrew.parker@gmail.com'
+
+
